Apply review changes
* Fix iterator resetting
