+++ dali/python/nvidia/dali/_utils/eager_utils.py
