+++ dali/test/python/test_eager_cpu_only.py
@@ -12,20 +12,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import json
 import numpy as np
 import os
-import tempfile
-import scipy.io.wavfile
 from functools import reduce
 
 from nvidia.dali import pipeline_def
 import nvidia.dali.experimental.eager as eager
 import nvidia.dali.fn as fn
-import nvidia.dali.tensors as tensors
 import nvidia.dali.types as types
-from test_audio_decoder_utils import generate_waveforms
 from test_utils import check_batch, get_dali_extra_path, get_files
+from test_dali_cpu_only_utils import *
 from webdataset_base import generate_temp_index_file as generate_temp_wds_index
 
 data_root = get_dali_extra_path()
@@ -54,8 +50,8 @@ def get_ops(op_path, fn_op=None, eager_op=None):
 
 
 @pipeline_def(batch_size=batch_size, num_threads=4, device_id=None)
-def no_input_pipeline(op, kwargs):
-    out = op(**kwargs)
+def reader_pipeline(op, kwargs):
+    out = op(pad_last_batch=True, **kwargs)
     if isinstance(out, list):
         out = tuple(out)
     return out
@@ -63,36 +59,25 @@ def no_input_pipeline(op, kwargs):
 
 def check_reader(op_path, *, fn_op=None, eager_op=None, batch_size=batch_size, N_iterations=2, **kwargs):
     fn_op, eager_op = get_ops(op_path, fn_op, eager_op)
-    pipe = no_input_pipeline(fn_op, kwargs)
+    pipe = reader_pipeline(fn_op, kwargs)
     pipe.build()
 
     iter_eager = eager_op(batch_size=batch_size, **kwargs)
 
     for _ in range(N_iterations):
-        for i, out2 in enumerate(iter_eager):
-            out1 = pipe.run()
+        for i, out_eager in enumerate(iter_eager):
+            out_fn = pipe.run()
 
-            if not isinstance(out2, (tuple, list)):
-                out2 = (out2,)
+            if not isinstance(out_eager, (tuple, list)):
+                out_eager = (out_eager,)
 
-            for o1, o2 in zip(out1, out2):
+            for tensor_out_fn, tensor_out_eager in zip(out_fn, out_eager):
                 if i == len(iter_eager) - 1:
-                    o1 = type(o1)([o1[j] for j in range(len(o2))])
-                out1_data = o1.as_cpu() if isinstance(o1, tensors.TensorListGPU) else o1
-                out2_data = o2.as_cpu() if isinstance(o2, tensors.TensorListGPU) else o2
+                    tensor_out_fn = type(tensor_out_fn)(
+                        [tensor_out_fn[j] for j in range(len(tensor_out_eager))])
 
-                check_batch(out1_data, out2_data, len(o2))
-
-
-@pipeline_def(batch_size=batch_size, num_threads=4, device_id=None)
-def reader_op_pipeline(op, kwargs, source=None, layout=None):
-    if source is None:
-        raise RuntimeError('No source for file reader.')
-    data, _ = fn.readers.file(file_root=source)
-    out = op(data, **kwargs)
-    if isinstance(out, list):
-        out = tuple(out)
-    return out
+                assert type(tensor_out_fn) == type(tensor_out_eager)
+                check_batch(tensor_out_fn, tensor_out_eager, len(tensor_out_eager))
 
 
 def test_file_reader_cpu():
@@ -128,62 +113,11 @@ def test_caffe2_reader_cpu():
 
 
 def test_nemo_asr_reader_cpu():
-    with tempfile.TemporaryDirectory() as tmp_dir:
-        def create_manifest_file(manifest_file, names, lengths, rates, texts):
-            assert(len(names) == len(lengths) == len(rates) == len(texts))
-            data = []
-            for idx in range(len(names)):
-                entry_i = {}
-                entry_i['audio_filepath'] = names[idx]
-                entry_i['duration'] = lengths[idx] * (1.0 / rates[idx])
-                entry_i['text'] = texts[idx]
-                data.append(entry_i)
-            with open(manifest_file, 'w') as f:
-                for entry in data:
-                    json.dump(entry, f)
-                    f.write('\n')
-        nemo_asr_manifest = os.path.join(tmp_dir, 'nemo_asr_manifest.json')
-        names = [
-            os.path.join(tmp_dir, 'dali_test_1C.wav'),
-            os.path.join(tmp_dir, 'dali_test_2C.wav'),
-            os.path.join(tmp_dir, 'dali_test_4C.wav')
-        ]
+    tmp_dir, nemo_asr_manifest = setup_test_nemo_asr_reader_cpu()
 
-        freqs = [
-            np.array([0.02]),
-            np.array([0.01, 0.012]),
-            np.array([0.01, 0.012, 0.013, 0.014])
-        ]
-        rates = [22050, 22050, 12347]
-        lengths = [10000, 54321, 12345]
-
-        def create_ref():
-            ref = []
-            for i in range(len(names)):
-                wave = generate_waveforms(lengths[i], freqs[i])
-                wave = (wave * 32767).round().astype(np.int16)
-                ref.append(wave)
-            return ref
-
-        ref_i = create_ref()
-
-        def create_wav_files():
-            for i in range(len(names)):
-                scipy.io.wavfile.write(names[i], rates[i], ref_i[i])
-
-        create_wav_files()
-
-        ref_text_literal = [
-            'dali test 1C',
-            'dali test 2C',
-            'dali test 4C',
-        ]
-        nemo_asr_manifest = os.path.join(tmp_dir, 'nemo_asr_manifest.json')
-        create_manifest_file(nemo_asr_manifest, names, lengths, rates, ref_text_literal)
-
-        fixed_seed = 1234
+    with tmp_dir:
         check_reader('readers.nemo_asr', manifest_filepaths=[nemo_asr_manifest], dtype=types.INT16,
-                     downmix=False, read_sample_rate=True, read_text=True, seed=fixed_seed)
+                     downmix=False, read_sample_rate=True, read_text=True, seed=1234)
 
 
 def test_video_reader():
@@ -192,21 +126,7 @@ def test_video_reader():
 
 
 def test_numpy_reader_cpu():
-    with tempfile.TemporaryDirectory() as test_data_root:
-        def create_numpy_file(filename, shape, typ, fortran_order):
-            arr = rng.random(shape) * 10.
-            arr = arr.astype(typ)
-            if fortran_order:
-                arr = np.asfortranarray(arr)
-            np.save(filename, arr)
-
-        num_samples = 20
-        filenames = []
-        for index in range(0, num_samples):
-            filename = os.path.join(test_data_root, 'test_{:02d}.npy'.format(index))
-            filenames.append(filename)
-            create_numpy_file(filename, (5, 2, 8), np.float32, False)
-
+    with setup_test_numpy_reader_cpu() as test_data_root:
         check_reader('readers.numpy', file_root=test_data_root)
